// src/lib/types.ts
/*
 Generated by typeshare 1.11.0
*/

export enum Role {
  User = 'User',
  Assistant = 'Assistant',
}

export interface Message {
  role: Role;
  content: string;
}

export enum Provider {
  OpenAI = 'OpenAI',
  Ollama = 'Ollama',
}

export interface Chat {
  id: string;
  name: string;
  messages: Message[];
  provider: Provider;
  model: string;
}